did we have a good reason for making MeasurementStorage an extension point?
seems like these 1D/2D plaintext output is pretty builtin at this point. @rainwoodman
